# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DID Resolver for did:cheqd method",
        "contact": {
            "name": "Cheqd Foundation Limited"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "resolver.cheqd.net",
    "basePath": "/1.0/identifiers",
    "paths": {
...
        "/{did}/resources/{resourceId}/metadata": {
            "get": {
                "description": "Get metadata for a specific Resource within a DID Resource Collection",
                "consumes": [
                    "application/did+ld+json",
                    "application/ld+json",
                    "application/did+json"
                ],
                "produces": [
                    "application/did+ld+json",
                    "application/ld+json",
                    "application/did+json"
                ],
                "tags": [
                    "Resource Resolution"
                ],
                "summary": "Fetch Resource-specific metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full DID with unique identifier",
                        "name": "did",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource-specific unique identifier",
                        "name": "resourceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DidDereferencing"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.IdentityError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.IdentityError"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/types.IdentityError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.IdentityError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.ContentType": {
            "type": "string",
            "enum": [
                "application/did+json",
                "application/did+ld+json",
                "application/ld+json",
                "application/json"
            ],
            "x-enum-varnames": [
                "DIDJSON",
                "DIDJSONLD",
                "JSONLD",
                "JSON"
            ]
        },
        "types.DereferencedResource": {
            "type": "object",
            "properties": {
                "checksum": {
                    "type": "string",
                    "example": "a95380f460e63ad939541a57aecbfd795fcd37c6d78ee86c885340e33a91b559"
                },
                "created": {
                    "type": "string",
                    "example": "2021-09-01T12:00:00Z"
                },
                "mediaType": {
                    "type": "string",
                    "example": "image/png"
                },
                "nextVersionId": {
                    "type": "string",
                    "example": "d4829ac7-4566-478c-a408-b44767eddadc"
                },
                "previousVersionId": {
                    "type": "string",
                    "example": "ad7a8442-3531-46eb-a024-53953ec6e4ff"
                },
                "resourceCollectionId": {
                    "type": "string",
                    "example": "55dbc8bf-fba3-4117-855c-1e0dc1d3bb47"
                },
                "resourceId": {
                    "type": "string",
                    "example": "398cee0a-efac-4643-9f4c-74c48c72a14b"
                },
                "resourceName": {
                    "type": "string",
                    "example": "Image Resource"
                },
                "resourceType": {
                    "type": "string",
                    "example": "Image"
                },
                "resourceURI": {
                    "type": "string",
                    "example": "did:canow:testnet:55dbc8bf-fba3-4117-855c-1e0dc1d3bb47/resources/398cee0a-efac-4643-9f4c-74c48c72a14b"
                }
            }
        },
        "types.DereferencingMetadata": {
            "type": "object",
            "properties": {
                "contentType": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ContentType"
                        }
                    ],
                    "example": "application/did+ld+json"
                },
                "did": {
                    "$ref": "#/definitions/types.DidProperties"
                },
                "error": {
                    "type": "string"
                },
                "retrieved": {
                    "type": "string",
                    "example": "2021-09-01T12:00:00Z"
                }
            }
        },
        "types.DidDereferencing": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "string",
                    "example": "https://w3id.org/did-resolution/v1"
                },
                "contentMetadata": {
                    "$ref": "#/definitions/types.ResolutionDidDocMetadata"
                },
                "contentStream": {},
                "dereferencingMetadata": {
                    "$ref": "#/definitions/types.DereferencingMetadata"
                }
            }
        },
...
        "types.DidProperties": {
            "type": "object",
            "properties": {
                "didString": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "methodSpecificId": {
                    "type": "string"
                }
            }
        },
...
        "types.IdentityError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "contentType": {
                    "$ref": "#/definitions/types.ContentType"
                },
                "did": {
                    "type": "string"
                },
                "internal": {},
                "isDereferencing": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "types.ResolutionDidDocMetadata": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string",
                    "example": "2021-09-01T12:00:00Z"
                },
                "deactivated": {
                    "type": "boolean",
                    "example": false
                },
                "linkedResourceMetadata": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.DereferencedResource"
                    }
                },
                "updated": {
                    "type": "string",
                    "example": "2021-09-10T12:00:00Z"
                },
                "versionId": {
                    "type": "string",
                    "example": "4979BAF49599FEF0BAD5ED0849FDD708156761EBBC8EBE78D0907F8BECC9CB2E"
                }
            }
        },
...
    }
}